app/requirements.txt
aiohttp
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary